requirements.txt
flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
python-telegram-bot==13.7
gunicorn==20.1.0
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
MarkupSafe==2.0.1
click==8.0.1
psycopg2-binary==2.9.1
pyOpenSSL==21.0.0
google-auth==2.15.0
google-auth-oauthlib==0.7.1
google-auth-httplib2==0.1.0
google-api-python-client==2.70.0
openai==1.3.0
numpy==1.21.0
scikit-learn==1.0.2
sentence-transformers==2.2.2
PyPDF2==3.0.1
Pillow==9.0.0
pytesseract==0.3.10
python-docx==0.8.11
cryptography==3.4.8
